dcal: handle all/none for positional
This is a temporary measure.

I dislike the usage:

    usage: dcal [-h] [day] [month] [year]

What I wanted was it to take a single `[day month year]` argument and
return it as a tuple/list.  This doesn't seem easy with `argparse`.  I
was able to set it to take all three arguments but I couldn't as easily
configure the help message or make the positional arguments optionial.